Created stype.cc
